+++ securedrop/source.py
@@ -72,19 +72,6 @@ def setup_g():
         g.loc = store.path(g.filesystem_id)
 
 
-@app.before_request
-@ignore_static
-def check_tor2web():
-    # ignore_static here so we only flash a single message warning
-    # about Tor2Web, corresponding to the initial page load.
-    if 'X-tor2web' in request.headers:
-        flash(Markup(gettext(
-            '<strong>WARNING:</strong> You appear to be using Tor2Web. '
-            'This <strong>does not</strong> provide anonymity. '
-            '<a href="/tor2web-warning">Why is this dangerous?</a>')),
-              "banner-warning")
-
-
 @app.route('/generate', methods=('GET', 'POST'))
 def generate():
     if logged_in():
